Use native brew-tap instead of cloning the tap repo manually
I decided to let library-repos to take the responsibility of auditing formulas instead of auditing them at the tap-repo when PRs are open.
And to utilize the built in `brew audit` command, given that the command can only be run on formulas it taps (as of now, 2021 Jan), we must let the change on the formula be visible in the tap.
I might as well do everything in the tap repo cloned by brew.